LPS-199470 Inline SiteMembershipPolicyFactoryImpl logic into Util
# breaking_change_report

## What portal-impl/src/com/liferay/portal/security/membershippolicy/SiteMembershipPolicyFactoryImpl.java

SiteMembershipPolicyFactoryImpl has been removed

## Why
It is for task 'LPS-190426', we would like to merge SiteMembershipPolicyFactoryImpl logic into SiteMembershipPolicyFactoryUtil

## Alternatives
We can use the static method in SiteMembershipPolicyFactoryUtil instead

----
